requirements.txt
crewai
crewai-tools
langchain
langchain-ollama
yfinance
streamlit
pandas
plotly
scipy
numpy
textblob
pandas-ta
praw 
torch
transformers
pyyaml
unstructured

crewai>=0.28.0
crewai-tools>=0.1.13
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.15
yfinance>=0.2.35
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.18.0
scipy>=1.12.0
numpy>=1.24.0
textblob>=0.17.1
pandas-ta>=0.3.14b0
praw>=7.7.1
torch>=2.1.0
transformers>=4.38.0
pyyaml>=6.0
python-dotenv>=1.0.0